requirements.txt
# ============================================
# REQUIREMENTS CHO DỰ ÁN TỐT NGHIỆP
# HỆ THỐNG HỖ TRỢ GIÁO VIÊN THCS DỰA TRÊN KNOWLEDGE GRAPH
# ============================================

# RDF và Knowledge Graph
rdflib==7.0.0
rdflib-jsonld==0.6.2

# Machine Learning
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Graph Processing
networkx==3.1

# Data Visualization (optional)
matplotlib==3.7.2
seaborn==0.12.2

# Web Framework (nếu cần backend API)
flask==2.3.2
flask-cors==4.0.0

# Database (nếu dùng GraphDB)
# graphdb-python==1.0.0  # Uncomment nếu cần

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

